aura-bot/requirements.txt
python-telegram-bot==21.9
anthropic>=0.40.0
aiohttp>=3.10.0
aiosqlite>=0.20.0
librouteros>=3.3.0
python-dotenv>=1.0.0
rapidfuzz>=3.6.0
